Refactor/code reuse (#231)
* feat: refactor gatewayConfig into small components

* refactor: remove old/depreciated AppConfig folder

* style: remove hardcoded need-webhook setup check

* feat: refactor all status-tag

* feat: refactor status-tag to cover: 'value not exist in enum' case
