--- src/components/settings/paymentGateways.css
@@ -1,5 +1,0 @@
-.payment-gateway-config-item:hover {
-  background-color: #f6f6f6;
-  border-radius: 4px;
-  cursor: grab;
-}
