+++ src/components/subscription/modals/changeSubStatus.tsx
@@ -4,7 +4,7 @@ import { Dayjs } from 'dayjs'
 import { useState } from 'react'
 import { markAsIncompleteReq } from '../../../requests'
 import { ISubscriptionType } from '../../../shared.types'
-import { SubscriptionStatus } from '../../ui/statusTag'
+import { SubscriptionStatusTag } from '../../ui/statusTag'
 
 interface Props {
   subInfo: ISubscriptionType | null
@@ -58,7 +58,7 @@ const Index = ({ subInfo, closeModal, refresh }: Props) => {
       <Row>
         <Col span={8}>
           <div className="flex h-full items-center justify-center">
-            Mark as &nbsp; {SubscriptionStatus(7)} until
+            Mark as &nbsp; {SubscriptionStatusTag(7)} until
           </div>
         </Col>
         <Col span={16}>
