+++ src/components/ui/couponPopover.tsx
@@ -4,7 +4,11 @@ import dayjs from 'dayjs'
 import { isEmpty } from 'lodash'
 import { useNavigate } from 'react-router-dom'
 import { showAmount } from '../../helpers'
-import { DiscountCode } from '../../shared.types'
+import {
+  DiscountCode,
+  DiscountCodeBillingType,
+  DiscountType
+} from '../../shared.types'
 import { getDiscountCodeStatusTagById } from '../ui/statusTag'
 
 const Index = ({ coupon }: { coupon?: DiscountCode }) => {
@@ -49,7 +53,9 @@ const Index = ({ coupon }: { coupon?: DiscountCode }) => {
               Billing type
             </Col>
             <Col span={14}>
-              {coupon.billingType === 1 ? 'One-time' : 'Recurring'}
+              {coupon.billingType === DiscountCodeBillingType.ONE_TIME
+                ? 'One-time'
+                : 'Recurring'}
             </Col>
           </Row>
           <Row>
@@ -57,7 +63,9 @@ const Index = ({ coupon }: { coupon?: DiscountCode }) => {
               Discount type
             </Col>
             <Col span={14}>
-              {coupon.discountType === 1 ? 'Percentage' : 'Fixed amount'}
+              {coupon.discountType === DiscountType.PERCENTAGE
+                ? 'Percentage'
+                : 'Fixed amount'}
             </Col>
           </Row>
           <Row>
